Make TS stricter and code compliant
Make all icons TypeScript and delete unused ones

Add global type definition for frontendlogger

Make code compliant with new TS config
