--- src/assets/images/Bin.js
@@ -1,16 +1,0 @@
-import * as React from 'react';
-
-export default function Bin(props) {
-    return (
-        <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 24 24" width={16} height={16} {...props}>
-            <path
-                fill="none"
-                stroke="#FF0000"
-                strokeLinecap="round"
-                strokeLinejoin="round"
-                strokeMiterlimit="10"
-                d="M3.516 3.5h16v20h-16zm4-3h8v3h-8zm-6.5 3h22M7.516 7v12m4-12v12m4-12v12"
-            />
-        </svg>
-    );
-}
