--- src/assets/images/icons/Book.js
@@ -1,20 +1,0 @@
-import React from 'react';
-
-function Book() {
-    return (
-        <svg xmlns="http://www.w3.org/2000/svg" width="66" height="71" fill="none" viewBox="0 0 66 71">
-            <path fill="#E7E9E9" d="M4.98 0h48.564l9.965 9.627v58.966H4.98V0z"></path>
-            <path
-                fill="#BA3A26"
-                fillRule="evenodd"
-                d="M62.224 8.424v59.174H8.868c-2.02 0-3.659-1.582-3.659-3.535 0-1.952 1.639-3.535 3.659-3.535h47.425V0H7.062C3.162 0 0 3.055 0 6.826v55.04C0 68.023 4.837 71 11.207 71H66V8.424h-3.776z"
-                clipRule="evenodd"
-            ></path>
-            <path fill="#FFD399" d="M14.943 43.322H44.83V48.136H14.943z"></path>
-            <path fill="#FFD399" d="M14.943 31.288H44.83V36.102000000000004H14.943z"></path>
-            <path fill="#FFD399" d="M14.943 19.254H44.83V24.068H14.943z"></path>
-        </svg>
-    );
-}
-
-export default Book;
