+++ src/assets/images/icons/BubbleChatInformation.tsx
@@ -1,0 +1,15 @@
+import React from 'react';
+
+function BubbleChatInformation() {
+    return (
+        <svg xmlns="http://www.w3.org/2000/svg" width="24" height="24" viewBox="0 0 24 24">
+            <g fill="none" stroke="#000" strokeLinejoin="round" strokeMiterlimit="10">
+                <path d="M23.5 10.759c0 5.183-5.148 9.384-11.5 9.384-1.418 0-2.775-.211-4.029-.594l-6.609 3.076 2.588-5.166C1.82 15.756.5 13.383.5 10.759.5 5.576 5.648 1.376 12 1.376s11.5 4.2 11.5 9.383z"></path>
+                <path strokeLinecap="round" d="M12.516 14.5v-6h-2m0 6h4"></path>
+                <path d="M12.5 5.5a.5.5 0 11-1.002-.002.5.5 0 011.002.002z"></path>
+            </g>
+        </svg>
+    );
+}
+
+export default BubbleChatInformation;
