--- src/assets/images/icons/Desklamp.js
@@ -1,36 +1,0 @@
-import React from 'react';
-
-function Desklamp() {
-    return (
-        <svg xmlns="http://www.w3.org/2000/svg" width="70" height="102" fill="none" viewBox="0 0 70 102">
-            <path
-                fill="#FFBD66"
-                fillRule="evenodd"
-                d="M61.355 19.475a5.633 5.633 0 00-7.835 1.371 5.602 5.602 0 001.375 7.815 5.633 5.633 0 007.835-1.371 5.602 5.602 0 00-1.375-7.815z"
-                clipRule="evenodd"
-            ></path>
-            <path
-                fill="#78706A"
-                d="M4.4 102c-1.05 0-1.9-.837-1.9-1.87 0-1.033.851-1.87 1.9-1.87h43.7c1.049 0 1.9.837 1.9 1.87 0 1.033-.851 1.87-1.9 1.87H4.4z"
-            ></path>
-            <path
-                fill="#78706A"
-                d="M7.855 48.546l21.388 49.617c.41.95-.049 2.044-1.024 2.444-.974.399-2.097-.047-2.506-.998L3.9 49.004a1.829 1.829 0 01.234-1.845L35.305 6.733a1.947 1.947 0 012.68-.37c.846.62 1.016 1.79.38 2.613l-30.51 39.57z"
-            ></path>
-            <path
-                fill="#78706A"
-                fillRule="evenodd"
-                d="M11.25 49.63a4.994 4.994 0 00-5-4.988c-2.762 0-5 2.234-5 4.988a4.994 4.994 0 005 4.987c2.762 0 5-2.233 5-4.987z"
-                clipRule="evenodd"
-            ></path>
-            <path
-                fill="#E7E9E9"
-                fillRule="evenodd"
-                d="M47.3 6.643l-5.811-4.058C37.645-.1 32.44.697 29.865 4.365c-2.575 3.669-1.546 8.819 2.298 11.504l5.81 4.057c-1.214 7.242 1.824 14.976 8.415 19.58L68.622 7.833C62.032 3.232 53.707 3.03 47.3 6.643z"
-                clipRule="evenodd"
-            ></path>
-        </svg>
-    );
-}
-
-export default Desklamp;
