--- src/assets/images/icons/Envelope.js
@@ -1,41 +1,0 @@
-import React from 'react';
-
-function Envelope() {
-    return (
-        <svg xmlns="http://www.w3.org/2000/svg" width="100" height="100" fill="none" viewBox="0 0 100 100">
-            <path
-                fill="#99C2E8"
-                fillRule="evenodd"
-                d="M50 100c27.614 0 50-22.386 50-50S77.614 0 50 0 0 22.386 0 50s22.386 50 50 50z"
-                clipRule="evenodd"
-            ></path>
-            <path
-                fill="#194D62"
-                fillRule="evenodd"
-                d="M80 47.88V64.9H21V47.88l29.5-19.855L80 47.88z"
-                clipRule="evenodd"
-            ></path>
-            <path fill="#E7E9E9" d="M31.158 21H70.491V65.25H31.158z"></path>
-            <path
-                fill="#669DB4"
-                fillRule="evenodd"
-                d="M21 47.694l17.7 13.074h23.6L80 47.694v29.78a2.177 2.177 0 01-2.177 2.178H23.177A2.177 2.177 0 0121 77.475V47.694z"
-                clipRule="evenodd"
-            ></path>
-            <path
-                fill="#337C9B"
-                fillRule="evenodd"
-                d="M38.7 60.338h23.6L80 73.317v.637c0 3.342-2.641 6.051-5.9 6.051H26.9c-3.258 0-5.9-2.71-5.9-6.051v-.637l17.7-12.979z"
-                clipRule="evenodd"
-            ></path>
-            <path
-                fill="#59514B"
-                fillRule="evenodd"
-                d="M39.684 28.375h4.425c.814 0 1.474.55 1.474 1.23 0 .678-.66 1.228-1.474 1.228h-4.425c-.815 0-1.475-.55-1.475-1.229s.66-1.229 1.474-1.229zM46.907 35.75h14.561c.731 0 1.324.55 1.324 1.23 0 .678-.593 1.228-1.324 1.228h-14.56c-.732 0-1.325-.55-1.325-1.229s.593-1.229 1.324-1.229zM39.745 43.125h21.51c.849 0 1.537.55 1.537 1.23 0 .678-.688 1.228-1.537 1.228h-21.51c-.849 0-1.536-.55-1.536-1.229s.687-1.229 1.536-1.229zM39.745 50.5h21.51c.849 0 1.537.55 1.537 1.23 0 .678-.688 1.228-1.537 1.228h-21.51c-.849 0-1.536-.55-1.536-1.229s.687-1.229 1.536-1.229z"
-                clipRule="evenodd"
-            ></path>
-        </svg>
-    );
-}
-
-export default Envelope;
