--- src/assets/images/icons/ExternalLink.js
@@ -1,16 +1,0 @@
-import React from 'react';
-
-function ExternalLink() {
-    return (
-        <svg xmlns="http://www.w3.org/2000/svg" width="16" height="16" fill="none" viewBox="0 0 16 16">
-            <path
-                fill="#0067C5"
-                fillRule="evenodd"
-                d="M8 2.667V4H1.333v10.667H12V8h1.333v8H0V2.667h8zM16 0v6h-1.333V2.276l-7.772 7.771-.942-.942 7.77-7.772H10V0h6z"
-                clipRule="evenodd"
-            ></path>
-        </svg>
-    );
-}
-
-export default ExternalLink;
