+++ src/assets/images/icons/QuestionActive.tsx
@@ -1,0 +1,16 @@
+import React from 'react';
+
+function QuestionActive() {
+    return (
+        <svg xmlns="http://www.w3.org/2000/svg" width="24" height="24" fill="none" viewBox="0 0 24 24">
+            <path
+                fill="#0067C5"
+                fillRule="evenodd"
+                d="M12 0C5.383 0 0 5.383 0 12s5.383 12 12 12c6.616 0 12-5.383 12-12S18.616 0 12 0zm0 19a1 1 0 11-.002-1.998A1 1 0 0112 19zm.5-3.5v-2.535a3.508 3.508 0 003-3.465C15.5 7.57 13.93 6 12 6c-1.931 0-3.5 1.57-3.5 3.5a.5.5 0 001 0C9.5 8.121 10.621 7 12 7c1.378 0 2.5 1.121 2.5 2.5S13.378 12 12 12a.5.5 0 00-.5.5v3a.5.5 0 001 0z"
+                clipRule="evenodd"
+            ></path>
+        </svg>
+    );
+}
+
+export default QuestionActive;
