+++ src/assets/images/icons/QuestionInactive.tsx
@@ -1,0 +1,16 @@
+import React from 'react';
+
+function QuestionInactive() {
+    return (
+        <svg xmlns="http://www.w3.org/2000/svg" width="24" height="25" fill="none" viewBox="0 0 24 25">
+            <path
+                fill="#0067C5"
+                fillRule="evenodd"
+                d="M0 12c0 6.617 5.383 12 12 12s12-5.383 12-12S18.617 0 12 0 0 5.383 0 12zm1.043 0C1.043 5.958 5.958 1.043 12 1.043c6.042 0 10.956 4.915 10.956 10.957 0 6.042-4.914 10.956-10.956 10.956S1.043 18.042 1.043 12zM12 16.174a.522.522 0 01-.522-.522v-3.13c0-.288.234-.522.522-.522a2.612 2.612 0 002.609-2.609A2.612 2.612 0 0012 6.783 2.612 2.612 0 009.391 9.39a.522.522 0 01-1.043 0A3.656 3.656 0 0112 5.74a3.656 3.656 0 013.652 3.652 3.66 3.66 0 01-3.13 3.616v2.645a.522.522 0 01-.522.522zm1.043 2.087a1.043 1.043 0 11-2.086 0 1.043 1.043 0 012.086 0z"
+                clipRule="evenodd"
+            ></path>
+        </svg>
+    );
+}
+
+export default QuestionInactive;
