--- src/assets/images/icons/View.js
@@ -1,43 +1,0 @@
-import React from 'react';
-
-function View() {
-    return (
-        <svg xmlns="http://www.w3.org/2000/svg" width="124" height="105" fill="none" viewBox="0 0 124 105">
-            <path fill="#C2EAF7" d="M0 0H98.167V98.598H0z" transform="translate(12.916 2.561)"></path>
-            <path
-                fill="#B5BD33"
-                fillRule="evenodd"
-                d="M12.916 71.584c40.063-6.328 35.81 20.218 98.167 6.587v23.153H12.916v-29.74z"
-                clipRule="evenodd"
-            ></path>
-            <path
-                fill="#C6C2BF"
-                fillRule="evenodd"
-                d="M64.583 2.561h-5.167v46.098h-46.5v5.122h46.5v46.097h5.167V53.781h46.5v-5.122h-46.5V2.56z"
-                clipRule="evenodd"
-            ></path>
-            <path
-                fill="#fff"
-                fillRule="evenodd"
-                d="M37.438 24.692a5.453 5.453 0 01-2.898-.829 3.914 3.914 0 01-7.747.154 2.348 2.348 0 01-2.777-3.513 5.48 5.48 0 115.04-8.345 3.915 3.915 0 017.586 1.633 5.478 5.478 0 11.795 10.9zM76.177 22.884c0-2.387-2.044-4.323-4.566-4.323-1.843 0-3.43 1.035-4.151 2.524a6.371 6.371 0 00-3.24-.877c-3.48 0-6.304 2.766-6.304 6.177 0 3.41 2.823 6.176 6.305 6.176a6.35 6.35 0 004.514-1.865 3.333 3.333 0 002.441 1.041c1.802 0 3.262-1.382 3.262-3.088 0-.73-.268-1.401-.715-1.93 1.458-.722 2.454-2.168 2.454-3.835z"
-                clipRule="evenodd"
-            ></path>
-            <path
-                fill="#C6C2BF"
-                fillRule="evenodd"
-                d="M108.5 97.278h-93V5.148h93v92.13zm5.167 0V2.574A2.579 2.579 0 00111.083 0H12.917a2.579 2.579 0 00-2.584 2.574v94.704h-7.75A2.579 2.579 0 000 99.852v2.574A2.578 2.578 0 002.583 105h118.834a2.579 2.579 0 002.583-2.574v-2.574a2.579 2.579 0 00-2.583-2.574h-7.75z"
-                clipRule="evenodd"
-            ></path>
-            <mask width="124" height="105" x="0" y="0" maskUnits="userSpaceOnUse">
-                <path
-                    fill="#fff"
-                    fillRule="evenodd"
-                    d="M108.5 97.278h-93V5.148h93v92.13zm5.167 0V2.574A2.579 2.579 0 00111.083 0H12.917a2.579 2.579 0 00-2.584 2.574v94.704h-7.75A2.579 2.579 0 000 99.852v2.574A2.578 2.578 0 002.583 105h118.834a2.579 2.579 0 002.583-2.574v-2.574a2.579 2.579 0 00-2.583-2.574h-7.75z"
-                    clipRule="evenodd"
-                ></path>
-            </mask>
-        </svg>
-    );
-}
-
-export default View;
