+++ src/assets/images/icons/ViewDesklampBook.tsx
@@ -1,0 +1,79 @@
+import React from 'react';
+
+function ViewDesklampBook() {
+    return (
+        <svg xmlns="http://www.w3.org/2000/svg" width="245" height="110" fill="none" viewBox="0 0 245 110">
+            <path fill="#C2EAF7" d="M0 0H98.167V98.598H0z" transform="translate(12.917 2.561)"></path>
+            <path
+                fill="#B5BD33"
+                fillRule="evenodd"
+                d="M12.917 71.584c40.063-6.328 35.81 20.218 98.167 6.587v23.153H12.917v-29.74z"
+                clipRule="evenodd"
+            ></path>
+            <path
+                fill="#C6C2BF"
+                fillRule="evenodd"
+                d="M64.584 2.561h-5.167v46.098h-46.5v5.122h46.5v46.097h5.167V53.781h46.5v-5.122h-46.5V2.56z"
+                clipRule="evenodd"
+            ></path>
+            <path
+                fill="#fff"
+                fillRule="evenodd"
+                d="M37.439 24.692a5.453 5.453 0 01-2.898-.829 3.914 3.914 0 01-7.747.154 2.348 2.348 0 01-2.777-3.513 5.48 5.48 0 115.04-8.345 3.915 3.915 0 017.586 1.633 5.478 5.478 0 11.795 10.9zM76.178 22.884c0-2.387-2.044-4.323-4.566-4.323-1.843 0-3.43 1.035-4.151 2.524a6.371 6.371 0 00-3.24-.877c-3.48 0-6.304 2.766-6.304 6.177 0 3.41 2.823 6.176 6.305 6.176a6.35 6.35 0 004.514-1.865 3.333 3.333 0 002.441 1.041c1.802 0 3.262-1.382 3.262-3.088 0-.73-.268-1.401-.715-1.93 1.458-.722 2.454-2.168 2.454-3.835z"
+                clipRule="evenodd"
+            ></path>
+            <path
+                fill="#C6C2BF"
+                fillRule="evenodd"
+                d="M108.5 97.278h-93V5.148h93v92.13zm5.167 0V2.574A2.579 2.579 0 00111.083 0H12.917a2.579 2.579 0 00-2.584 2.574v94.704h-7.75A2.579 2.579 0 000 99.852v2.574A2.578 2.578 0 002.583 105h118.834a2.579 2.579 0 002.583-2.574v-2.574a2.579 2.579 0 00-2.583-2.574h-7.75z"
+                clipRule="evenodd"
+            ></path>
+            <mask width="124" height="105" x="0" y="0" maskUnits="userSpaceOnUse">
+                <path
+                    fill="#fff"
+                    fillRule="evenodd"
+                    d="M108.5 97.278h-93V5.148h93v92.13zm5.167 0V2.574A2.579 2.579 0 00111.083 0H12.917a2.579 2.579 0 00-2.584 2.574v94.704h-7.75A2.579 2.579 0 000 99.852v2.574A2.578 2.578 0 002.583 105h118.834a2.579 2.579 0 002.583-2.574v-2.574a2.579 2.579 0 00-2.583-2.574h-7.75z"
+                    clipRule="evenodd"
+                ></path>
+            </mask>
+            <path fill="#E7E9E9" d="M183.981 39h48.564l9.965 9.627v58.966h-58.529V39z"></path>
+            <path
+                fill="#BA3A26"
+                fillRule="evenodd"
+                d="M241.224 47.424v59.174H187.868c-2.02 0-3.659-1.582-3.659-3.536 0-1.951 1.639-3.534 3.659-3.534h47.425V39h-49.231c-3.9 0-7.062 3.055-7.062 6.826v55.04c0 6.157 4.837 9.134 11.208 9.134H245V47.424h-3.776z"
+                clipRule="evenodd"
+            ></path>
+            <path fill="#FFD399" d="M193.943 82.322H223.83V87.136H193.943z"></path>
+            <path fill="#FFD399" d="M193.943 70.288H223.83V75.102H193.943z"></path>
+            <path fill="#FFD399" d="M193.943 58.254H223.83V63.068H193.943z"></path>
+            <path
+                fill="#FFBD66"
+                fillRule="evenodd"
+                d="M156.355 27.475a5.634 5.634 0 00-7.835 1.371 5.602 5.602 0 001.375 7.815 5.634 5.634 0 007.835-1.371 5.602 5.602 0 00-1.375-7.815"
+                clipRule="evenodd"
+            ></path>
+            <path
+                fill="#78706A"
+                d="M99.4 110c-1.049 0-1.9-.837-1.9-1.87 0-1.033.851-1.871 1.9-1.871H143.1c1.05 0 1.901.838 1.901 1.871s-.851 1.87-1.901 1.87H99.401z"
+            ></path>
+            <path
+                fill="#78706A"
+                d="M102.855 56.546l21.388 49.617c.409.95-.049 2.044-1.024 2.444-.974.399-2.097-.047-2.506-.998L98.9 57.004a1.828 1.828 0 01.234-1.845l31.171-40.426a1.947 1.947 0 012.681-.37c.845.62 1.015 1.79.38 2.614l-30.511 39.569z"
+            ></path>
+            <path
+                fill="#78706A"
+                fillRule="evenodd"
+                d="M106.25 57.63a4.994 4.994 0 00-4.999-4.988 4.994 4.994 0 00-5.001 4.988 4.994 4.994 0 005.001 4.987 4.994 4.994 0 004.999-4.987"
+                clipRule="evenodd"
+            ></path>
+            <path
+                fill="#E7E9E9"
+                fillRule="evenodd"
+                d="M142.299 14.643l-5.81-4.058c-3.844-2.685-9.049-1.888-11.624 1.78-2.575 3.669-1.546 8.819 2.298 11.504l5.81 4.057c-1.214 7.242 1.824 14.976 8.415 19.58l22.234-31.672c-6.591-4.602-14.915-4.805-21.323-1.191z"
+                clipRule="evenodd"
+            ></path>
+        </svg>
+    );
+}
+
+export default ViewDesklampBook;
